+++ zython/operations/constraint.py
@@ -27,25 +27,30 @@ class Constraint:
         return Constraint(_Op_code.and_, self, other)
 
 
-def _exists(seq: "zython.var_par.types.ZnSequence",
-            iter_var: Optional["zython.var"] = None,
-            func: Optional[Union["Constraint", Callable]] = None) -> "Constraint":
+def _exists(
+    seq: "zython.var_par.types.ZnSequence",
+    iter_var: Optional["zython.var"] = None,
+    func: Optional[Union["Constraint", Callable]] = None,
+) -> "Constraint":
     return Constraint(_Op_code.exists, seq, iter_var, func)
 
 
-def _forall(seq: "zython.var_par.types.ZnSequence",
-            iter_var: Optional["zython.var"] = None,
-            func: Optional[Union["Constraint", Callable]] = None) -> "Constraint":
+def _forall(
+    seq: "zython.var_par.types.ZnSequence",
+    iter_var: Optional["zython.var"] = None,
+    func: Optional[Union["Constraint", Callable]] = None,
+) -> "Constraint":
     return Constraint(_Op_code.forall, seq, iter_var, func)
 
 
-def cumulative(start_times: "zython.var_par.types.ZnSequence",
-               durations: "zython.var_par.types.ZnSequence",
-               requirements: "zython.var_par.types.ZnSequence",
-               limit: int) -> "Constraint":
+def cumulative(
+    start_times: "zython.var_par.types.ZnSequence",
+    durations: "zython.var_par.types.ZnSequence",
+    requirements: "zython.var_par.types.ZnSequence",
+    limit: int,
+) -> "Constraint":
     return Constraint(_Op_code.cumulative, start_times, durations, requirements, limit)
 
 
-def table(x: "zython.var_par.types.ZnSequence",
-          t: "zython.var_par.types.ZnSequence") -> "Constraint":
+def table(x: "zython.var_par.types.ZnSequence", t: "zython.var_par.types.ZnSequence") -> "Constraint":
     return Constraint(_Op_code.table, x, t)